Rewrite ReadBitsInt{Be,Le}(), change param type from `uint8` to `int`
See https://github.com/kaitai-io/kaitai_struct/issues/949

Changing the parameter type is potentially a breaking change, because
calls like ReadBitsIntBe(uint8(15)) will no longer work. But this
will not be a problem for KSC-generated code, it could only break
manually-written (or manually-patched) code.

`ReadBitsInt(n uint8)` is intentionally left as it was, since it exists
only to maintain backward compatibility and will be removed at some point.

+++ kaitai/stream_test.go
@@ -721,7 +721,7 @@ func TestStream_AlignToByte(t *testing.T) {
 
 func TestStream_ReadBitsIntBe(t *testing.T) {
 	type args struct {
-		totalBitsNeeded uint8
+		totalBitsNeeded int
 	}
 	tests := []struct {
 		name    string
@@ -771,7 +771,7 @@ func TestStream_ReadBitsArray(t *testing.T) {
 
 func TestStream_ReadBitsIntLe(t *testing.T) {
 	type args struct {
-		n uint8
+		n int
 	}
 	tests := []struct {
 		name    string
